extract shared AddressRow model and row builder
